replacing uint32_t with size_t to reduce conversions

+++ kdtree.h
@@ -12,10 +12,9 @@ private:
     Node* _root;
     uint32_t _nodes; // total amount of nodes
     uint32_t _depth; // maximum depth
-    //uint32_t _iterated; // number of iterated nodes for raycast
 
 public:
-    KDTree (float *vertices, uint32_t size);
+    KDTree (float *vertices, size_t size);
 
     void print ();
     Node* root ();
