+++ vis.cpp
@@ -9,7 +9,7 @@
 
 float lastX, lastY = 0;
 
-Vis::Vis (KDTree &tree, float* vertices, uint32_t countVertices)
+Vis::Vis (KDTree &tree, float* vertices, size_t countVertices)
 	: _tree(tree), _vertices(vertices), _countVertices(countVertices), _showGrid(false)
 {
 	init();
@@ -48,9 +48,9 @@ int Vis::init ()
 	glBindVertexArray(VAO);
 
 	glBindBuffer(GL_ARRAY_BUFFER, VBO);
-	glBufferData(GL_ARRAY_BUFFER, sizeof(float) * _countVertices, _vertices, GL_STATIC_DRAW);
+	glBufferData(GL_ARRAY_BUFFER, sizeof(float)*_countVertices, _vertices, GL_STATIC_DRAW);
 
-	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(float), (void*)0);
+	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 3*sizeof(float), (void*)0);
 	glEnableVertexAttribArray(0);
 
 	static constexpr float planepoints[12] =
